backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.9.2
